Initialize shared parameters in constructor.
Finding it useful to be able to specify the parameters passed to a
shared function in the constructor when using Sequester to create a
latch. The use case is when a latch is sometimes necessary but not
always necessary, and when it is not always necessary you create a dummy
latch that is already open, but in order to do so you must provide the
value to the constructor that would have been provided by exclusive
action of the latch.

Closes #24.